requirements.txt
django==1.3
django-userena==1.0.1
Pillow
django-extensions
-e git+https://github.com/gsiegman/django-qrcodes.git#egg=django-qrcodes
